+++ internal/cli/commands/startup/cmd.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 // Package startup implements the startup command.
 package startup